requirements.txt
absl-py==2.3.1
altair==4.2.2
asttokens==3.0.0
astunparse==1.6.3
attrs==25.3.0
blinker==1.9.0
cachetools==5.5.2
certifi==2025.8.3
charset-normalizer==3.4.3
click==8.3.0
colorama==0.4.6
comm==0.2.3
contourpy==1.3.2
cycler==0.12.1
debugpy==1.8.17
decorator==5.2.1
entrypoints==0.4
et_xmlfile==2.0.0
exceptiongroup==1.3.0
executing==2.2.1
flatbuffers==25.2.10
fonttools==4.60.0
gast==0.4.0
gitdb==4.0.12
GitPython==3.1.45
google-auth==2.40.3
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
grpcio==1.75.0
h5py==3.8.0
idna==3.10
importlib_metadata==8.7.0
ipykernel==6.30.1
ipython==8.37.0
jax==0.4.30
jaxlib==0.4.30
jedi==0.19.2
Jinja2==3.1.6
joblib==1.5.2
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
jupyter_client==8.6.3
jupyter_core==5.8.1
keras==2.12.0
kiwisolver==1.4.9
libclang==18.1.1
Markdown==3.9
markdown-it-py==4.0.0
MarkupSafe==3.0.2
matplotlib==3.7.1
matplotlib-inline==0.1.7
mdurl==0.1.2
ml_dtypes==0.5.3
narwhals==2.5.0
nest-asyncio==1.6.0
numpy==1.23.5
oauthlib==3.3.1
openpyxl==3.1.5
opt_einsum==3.4.0
packaging==25.0
pandas==1.5.3
parso==0.8.5
Pillow==9.5.0
platformdirs==4.4.0
plotly==6.3.0
prompt_toolkit==3.0.52
protobuf==3.20.3
psutil==7.1.0
pure_eval==0.2.3
pyarrow==21.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydeck==0.9.1
Pygments==2.19.2
Pympler==1.1
pyparsing==3.2.5
python-dateutil==2.9.0.post0
pytz==2025.2
pywin32==311
pyzmq==27.1.0
referencing==0.36.2
requests==2.31.0
requests-oauthlib==2.0.0
rich==14.1.0
rpds-py==0.27.1
rsa==4.9.1
scikit-learn==1.2.2
scipy==1.15.3
seaborn==0.12.2
six==1.17.0
smmap==5.0.2
stack-data==0.6.3
streamlit==1.22.0
tenacity==8.5.0
tensorboard==2.12.3
tensorboard-data-server==0.7.2
tensorflow==2.12.0
tensorflow-estimator==2.12.0
tensorflow-intel==2.12.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==3.1.0
threadpoolctl==3.6.0
toml==0.10.2
toolz==1.0.0
tornado==6.5.2
traitlets==5.14.3
typing_extensions==4.15.0
tzdata==2025.2
tzlocal==5.3.1
urllib3==2.5.0
validators==0.35.0
watchdog==6.0.0
wcwidth==0.2.14
Werkzeug==3.1.3
wrapt==1.14.2
zipp==3.23.0
